GH-410: Ensure EOF gets sent before CLOSE when channel closes
When a channel using a ChannelAsyncOutputStream, do not only wait until
all pending data has been written but also wait for the EOF having been
sent before proceeding with closing the channel.

For channels using a synchronous ChannelOutputStream, no local fix was
found. Instead give the channel access to the IoWriteFuture for sending
the EOF packet, and if one was sent, ensure that closing the channel
proceeds only after the EOF has been sent.

Otherwise there may be a race between sending that EOF and sending the
CLOSE, and sometimes the EOF might be sent after the CLOSE.

Bug: https://github.com/apache/mina-sshd/issues/410