Fix race conditions in test
Some tests shut down the SshClient before closing SshSessions. This is
fine, but it leads to sessions being closed asynchronously. The client
initiates shutdown by closing the underlying IO framework, which is
asynchronous. When the IO channel (i.e., the network connection) closes,
it notifies the SshSession about it, which then closes itself. But this
occurs on an I/O thread, so tests cannot simply synchronously test for
the "session closed" event having been signalled.

Make the tests wait a little while on a CountDownLatch until the event
has been sent.

Also ensure thread safety in SftpServerTest.