GH-434: skip unknown public keys from external sources
Change Buffer.getPublicKey() to set the read position to after the key
even if reading the key fails. This enables us to continue reading keys
from a list of keys in a buffer even when a particular key cannot be
decoded.

Change the two places where we receive lists of public keys from
external sources: from an SSH agent or via the "hostkeys-00@openssh.com"
extension. Skip and log keys that cannot be decoded.