AbstractClientChannel: fix waiting for channel having closed
channel.waitFor(EnumSet.of(ClientChannelEvent.CLOSED), timeout) would
not wait until the channel was closed, but would return with the CLOSED
event already when the channel was still closing.

The closeSignaled flag was set already in preClose(), but at that point
the SSH_MSG_CHANNEL_CLOSE handshake still had to be done. In the common
usage of
```
try  (ClientSession session = ...) {
  // Authenticate, then:
  try (ChannelExec channel = session.createExecChannel(...)) {
    // Open the channel, perhaps do something with input/output streams
    // ...
    Collection<ClientChannelEvent> result =
        channel.waitFor(numSet.of(ClientChannelEvent.CLOSED), timeout);
  }
}
```
the call to waitFor() would return early already when the server's
SSH_MSG_CHANNEL_CLOSE was received, but before the client's reply was
sent.

The thread executing this code would then close the channel and then
the session, which results in a race with sending that client reply out.
If the session got closed first, the client would log an exception about
a write attempt on a closed session.

Fix this by removing closeSignaled.get() from the conditions for being
in CLOSED state. The state is not CLOSED yet; it would be IS_CLOSING at
best.