Switch uses of JSch library to the com.github.mwiede:jsch fork
The original Jcraft version of the JSch library has not seen updates in
over a decade. The current fork with the most active maintenance is the
com.github.mwiede:jsch library and it is largely compatible with the
original but with secure defaults. It also absorbed the Jzlib
implementation in its own code base and eliminated the need to have the
extra dependencies.

As this is only a test dependency for this code base I feel adding this
update is an improvement.