Improve javadoc on ClientChannel.setIn(InputStream)
Closing that input stream when the channel closes is in my view a design
mistake. The stream is obtained by the application code from somewhere,
and it should be the application's responsibility to close it when
appropriate. While changing the implementation (not closing the stream
in AbstractClientChannel.getInnerCloseable()) would work, it appears to
be too risky a change for a minor version bump.

So just explain in the javadoc what to do if the stream should _not_ be
closed. (For instance, System.in should never be closed and doing so may
even cause hangs on Windows.)

Also add a (somewhat contrived) test that chains two remote shells by
setting the output of one shell as the input of another shell.