sftp readdir: determine file type from longname
Some SFTP v3 servers do not include the file type flags in the
permissions field of an SSH_FXP_NAME record. It this case use the
"longname" field to extract this information, if possible.

Also give the SftpClientDirectoryScanner and the DirectoryScanner a
flag to make them return not only regular files but also links and
other items. (DirectoryScanner already returned links to regular files;
SftpClientDirectoryScanner did not.)

Bug: https://github.com/apache/mina-sshd/issues/489