GH-493: Fix "blocksize" of arcfour ciphers
Arcfour is a stream cipher. The padding must then be such that the total
SSH packet size is a multiple of 8. See RFC 4253: "the length of the
concatenation of 'packet_length', 'padding_length', 'payload', and
'random padding' MUST be a multiple of the cipher block size or 8,
whichever is larger. This constraint MUST be enforced, even when using
stream ciphers."[1]

Previous code had 16 and 32 as "block size" for these ciphers, which is
wrong. Set it to 8.

The wrong values 16 and 32 were added by mistake in commit 1166a5d2 in
version 2.2.0.

[1] https://datatracker.ietf.org/doc/html/rfc4253#page-7

Bug: https://github.com/apache/mina-sshd/issues/493