GH-509: AbstractSftpClient: validate SSH_FXP_ATTRS flags
In SFTP versions >= 4, the flags must not include flag 0x2
(SSH_FILEXFER_ATTR_UIDGID).[1] Throw an exception if the client
receives this flag from the server, and ensure the client does
not send this flag.

On the server side, we never send back this flag in SFTP >= v4,
and we silently ignore it if a client sends it.

[1] https://datatracker.ietf.org/doc/html/draft-ietf-secsh-filexfer-03#section-5.1

Bug: https://github.com/apache/mina-sshd/issues/509