Harden CBC cipher handling
Include the mitigations implemented in OpenSSH for a particular kind of
attack against CBC ciphers.[1][2]

The attack is thwarted by making disconnections due to a bad packet
length or due to a failing MAC check indistinguishable. When we detect
a bad packet length, we substitute a random packet length instead and
keep on requesting data, so an attacker cannot learn anything about the
initial 32bits of a packet. We then continue normally, until the MAC
check fails. In the very unlikely case that it doesn't fail, we
disconnect unconditionally after the MAC check. The disconnection
messages are identical in all these cases.

A note on the commit message of commit edb2b951: the attack described
in [2] is feasible only within one and the same SSH connection, where it
has a probability of 2**-14 or 2**-18 to learn the first 14 or 32 bits
of a captured 16-byte encrypted block that is re-injected later in that
same connection. (I had originally misunderstood their description.) If
that block is injected in a different SSH connection with different
keys, nothing is revealed. As such it makes sense to include a
mitigation for this even if CBC modes are not recommended in SSH, since
we do still provide the implementation to enable users to connect to
legacy systems.

The authors of [2] later published a follow-up paper [3] in which they
claimed the OpenSSH mitigation was still vulnerable via a timing side
channel. This timing weakness should not exist in the mitigation
included here.

[1] https://www.openssh.com/txt/cbc.adv
[2] https://www.cs.umd.edu/~jkatz/security/downloads/PlaintextRecoverySSH.pdf
[3] https://pure.royalholloway.ac.uk/ws/portalfiles/portal/27577238/surfeit.pdf