GH-524: Optimize the chacha20-poly1305 cipher
Unroll the permutation loop of ChaCha20 and simplify the crypt() method
to work on bytes instead of ints. Unroll and optimize the pack/unpack
methods in ChaCha20 and in Poly1305. Don't use Integer.toUnsignedLong():
avoid the extra method call by doing (int & 0xFFFF_FFFFL) inline. In
Poly1305.processBlock(), use ints instead of longs where possible
(variables t0-t3).

All this brings a speed-up of about 40% for the encryption/decryption.
Of the time spent in ChaCha20-Poly1305 about one third is spent in the
ChaChaEngine and two thirds in Poly1305.processBlock().

Bug: https://github.com/apache/mina-sshd/issues/524