Fix tests using EOL CentOS 7.9 image
It appears that the yum repository has gone. So move away from CentOS
images for tests against old OpenSSH versions.

Remove SessionsReKeyJostExchangeTest. It tested SSHD-1264, which could
never reproduced with other containers, not even with an Alpine using
an older OpenSSH. Finding an alternative for the CentOS image on which
SSHD-1264 would be reproducible would require testing first with an old
Apache MINA sshd in which SSHD-1264 wasn't fixed yet. I don't have the
time for this.

Remove ArcFourOpenSShTest. This test was useful to analyze a problem
and verify a fix, but the test took a lot of time and was completely
overblown. Replaced by a simple JUnit test method that verifies that all
ciphers have block sizes >= 8.

For the StrictKexInteroperabilityTest, use alpine:3.9.2 instead of
CentOS. That old Alpine installs OpenSSH 7.9, which also never got the
"strict kex" fixes.