GH-524: Performance: avoid buffer copying
In ChannelAsyncOutputStream, do not create a new SSH packet if the
caller already had passed in an SSH packet buffer, and we're writing
it fully. This avoids allocating a new buffer, which gives GC less
to do, and it also avoid needlessly copying the data around. Instead
the prepared SSH packet buffer is encoded and written directly.

Bug: https://github.com/apache/mina-sshd/issues/524