GH-524: Use built-in native AES from SunJCE if available
Bouncy Castle apparently only has native implementations in its "LTS"
releases. BC 1.78.1 has none. SunJCE uses native code.

The "security registrar" architecture in Apache MINA sshd prefers
registered providers over the platform providers, and it automatically
registers Bouncy Castle if it's present. So with Bouncy Castle on the
classpath and an SSH connection for which an AES cipher was specified,
Apache MINA sshd would use the much slower Bouncy Castle AES.

Add a new "SunJCEWrapper" registrar that brings the SunJCE AES and
HmacSHA* to the front of the list, so that they are used even if Bouncy
Castle is also registered.

The new registrar can be disabled via a system property as usual, and
it is only enabled if the JVM indeed has a "SunJCE" security provider
registered.

See also https://issues.apache.org/jira/browse/SSHD-719 (issue closed
as "not needed").

Bug: https://github.com/apache/mina-sshd/issues/524