GH-524: Add sshd-benchmarks maven project
The new maven project sshd-benchmarks contains JMH benchmarks. The
project is not part of the binary distribution, and the install:install
and deploy:deploy targets are skipped. The project is included in the
source distribution.

The benchmarks are intended to be run locally.

Currently the project contains benchmarks for SFTP file uploads, which
can be run either against a local container (if the docker engine is
running), or against an external SFTP server.