GH-636: Handle unknown key types in known_hosts
Harden the parser so that it can parse known_host and authorized_key
lines with unknown key types. Introduce a new UnsupportedSshPublicKey
class to be able to deal with such entries later on when the server
host key is compared. (An alternative would have been not to create
PublicKeys from known_host lines at all but to serialize the given
server key into string form and then just compare against the string
from the known_host line. But that is not possible without breaking
API...)

Such an UnsupportedSshPublicKey supports getting its key type, its raw
key data, its fingerprint, and it can be written into a Buffer.