KEX: by default abort on invalid certificates
Switch the default of CoreModuleProperties.ABORT_ON_INVALID_CERTIFICATE
from false to true.

Unless we get tons of bug reports due to this I plan to deprecate and
remove this setting altogether. IMO an SSH client should *never* accept
connections if a certificate key exchange is used and the certificate
the server sent is invalid. (Expired, wrongly signed, not listing the
host name or IP in the principals, or whatever.)