[SSHD-1167] KEX: handle "@cert-authority" in known_hosts files
Implement the OpenSSH behavior for checking of certificates: if the
server key is a certificate, only entries with markers @revoked or
@cert-authority are considered. There must be an entry marked as
@cert-authority matching the certificate's CA public key, otherwise the
server key is rejected. Unknown CA keys are never added to the file.

Move the responsibility for this check from AbstractClientSession to
the ServerKeyVerifier. Note that this means that the provided
implementation does *not* fall back to checking the certificate's
public key (i.e., the one that was signed by the CA key). If users
want that (insecure) behavior, they have to provide their own
ServerKeyVerifier that implements it.