[SSHD-1161] Pubkey auth: handle certificates in PublickeyAuthenticator
Move the responsibility of checking the user name against the
certificate's principals into the PublickeyAuthenticator.

First, if not, an "accept all" authenticator wouldn't work since the
session might have rejected the certificate already.

Second, the OpenSSH format allows the user to explicitly specify
principals in the authorized_keys file. If so, these principals override
the log-in user name.