[SSHD-1161] Pubkey auth: source-address restriction in certificates
OpenSSH certificates may contain a critical option "source-address". If
present, it contains a comma-separated list of CIDR ranges from where
the certificate is accepted for authentication.[1]

Extract and parse this option, and do not accept the certificate if the
client's IP address is not contained in at least one CIDR range given.

[1] https://github.com/openssh/openssh-portable/blob/master/PROTOCOL.certkeys