GH-663: Fix a race in IoSession creation
IoSessions are created asynchronously through an IoConnector. When an
IoConnector is closed, all sessions created through it are also to be
closed.

Because session creation is asynchronous, it was possible that a newly
created session registered on an already closed IoConnector, and then
would never be closed.

Prevent this by forcibly closing any newly created IoSession if the
IoConnector is already closed when the new session tries to register
with the connector.

Add a test that verifies that the connect future the client code sees
does not provide a session but is fulfilled either by an exception or
by having been cancelled.