GH-690: Handle append mode for buggy SFTP v3 servers
Some SFTP v3 servers have trouble with append mode for file uploads.
OpenSSH < 6.3 didn't have append mode implemented, and some server
identifying as "7.9.0.0_openssh" apparently also cannot do it.

As a work-around get the file size first, then use that as base offset
for writing. This is at best an approximation of true append mode. We
still send the "append" flag, so for correct servers there will be no
change, but with the work-around appending may work as expected in most
cases even for servers that get append mode wrong.