GH-677: Fix ScpShell
Fix the handling of the home directory when a RootedFileSystem is used.

'currentDir' and 'homeDir' must be paths from the file system used.
WinSCP may, at least initially, send Windows-style paths; so ensure
that the path is converted.

When listing the root directory, don't include the ".." entry for the
parent directory.

Tested manually on Windows.