Improve packet padding
Don't add random extra padding to low-level messages.

OTOH ensure that the padding for messages that might carry user
passwords is always at least 64 bytes to prevent that anyone could make
guesses about user password lengths (if an attacker can identify these
messages at all).

Change the outgoing interfaces to pass through the SSH command byte
together with the buffer. Otherwise the CryptFilter, which decides on
the amount of padding, may not have access to it if compression is used.
(The command byte would also be compressed and thus cannot be read
anymore from the buffer directly.)