Support for customizing the filter chain
Simplify calling the SessionListener, and add a new listener callback
sessionStarting(). A session listener thus has two possibilities to
override the default filter chain: either define the whole chain in
sessionCreated(), or add filters to the default chain in
sessionStarting().

Add tests for client and server verifying that a custom filter added in
sessionStarting() is called.