Server-side HAProxyProtocolFilter
Provide an IoFilter that can be added at the front of the filter chain
of a server session to handle the HAProxy protocol header. The filter
handles both protocol versions 1 and 2.

Remove all the ServerProxyAcceptor handling in the server session, and
completely drop this interface. The session doesn't need to know about
this anymore; user code can just add the filter and have this handled
transparently for the session.

Move and rewrite the existing ServerProxyAcceptorTest.