Simplify port forwarding listeners
Reduce code duplication. Introduce an IOConsumer, and make the listener
invocations throw IOException instead of Throwable (which is bad
practice anyway).

Also, the code was prepared to walk through a hierarchy of
PortEventListenerManagerHolders. But the only classes implementing this
interface are ClientConnectionService and ServerConnectionService. So
there is no hierarchy. The AbstractConnectionServiceFactory and the
Session register themselves in the connection service.