Remove ClientProxyConnector
This interface was never useful; on the client side, connecting through
a proxy to an SSH server is always a request-reply protocol, not a
single message.

With the new filter chain, connecting through a proxy can be implemented
with a filter. The SshClient will need some changes, too, since it must
open the connection not to the SSH server but to the proxy server. A
full implementation is deferred to later.