Move SSH_MSG_GLOBAL_REQUEST handling to the ConnectionService
Now that all the SSH transport stuff has gone from AbstractSession,
it's evident that the session contains something it shouldn't: global
requests are a feature of the connection service. So move the whole
implementation into the AbstractConnectionService. Keep the convenience
request() methods on the session, but just forward to the service, if
the current service _is_ a ConnectionService.