GH-727: supply default port for proxyJump if no HostConfigEntry
If SshClient creates an empty HostConfigEntry, supply the default port
if the given port is <= 0.

I'd have liked to do that once and for all in HostConfigEntry.getPort(),
but HostConfigEntry.append(Appendable a) relies on getPort() returning a
value <= 0 if not set.

If there _is_ a HostConfigEntry for the proxyJump, then
HostConfigEntryResolver.resolveEffectiveHost() would already supply the
default port. (That bit was fixed in version 2.10.)