GH-729: client-side proxy support
Bring over the client-side proxy support from JGit and adapt it for use
as a ClientProxyFilter in the filter chain of an AbstractClientSession.

Includes integration tests for HTTP or SOCKS proxies. Also test using a
dynamic port forwarding, and fix the SOCKS listener in Apache MINA sshd.

Also change Sshd1033Test to use ephemeral ports instead of fixed ports.