GH-733: Fix SftpRemotePathChannel.transferTo
Give the SftpInputStreamAsync a logical file size larger than the
current position transfer is requested to start at. If the user
calls transferTo(position, count, output), the logical file size for
the SftpInputStreamAsync should be (position + count). It's no problem
if that is larger than the real file size, it'll just produce an EOF
return.

Also guard against NPE in SftpInputStreamAsync in case the initial
position is beyond the alleged file size.