Add API comparison reports
Create japicmp reports, collect them in a web site, and publish that web
site via gh-pages.

Because the maven-site-plugin generated rather useless stuff for the
top-level multi-module project (maybe I just don't know how to use it),
I decided to create that web site in a small sshd-site module, copying
over the japicmp reports explicitly and then generating a site for this
simple maven project.

Using groovy, we automatically adapt the site menu, and we also enhance
the HTML output a little: japicmp's markdown output makes use of the
GitHub markdown "alert" extension, which Doxia (the markdown-to-html
converter the maven-site-plugin uses) doesn't know. But it's fairly easy
to post-process the HTML output to render these alerts properly.

The web site is published using maven-scm-publish-plugin to gh-pages. It
is accessible at https://apache.github.io/mina-sshd .