Simplify CoreModuleProperties.SEND_IMMEDIATE_KEXINIT
Setting this property to false should not only delay sending the
client's initial KEX_INIT until has received the server. It should
delay the whole KEX, including determining the client's proposal,
until that point, so that the client can adapt its own proposal
based on what the server sent (ident and server proposal).

So remove the DelayKexFilter, and just don't do anything in
KexFilter.startKex() if the property is false and it's the very first
call. (Well, we send a null message through the filter chain to tell
lower filters that they should initialize. In particular, this triggers
the IdentFilter to send the client's SSH protocol version string.)

Add a test to verify that we do have the server's ident and proposal
when the client finally builds its own proposal.