[releng] Work around impsort plugin problem
When building on windows, the impsort plugin gets thrown off by the
comment just before the imports in BCrypt.java if the file as checked
out uses plain LF line endings.

It somehow insists on changing the 2nd of the three lines by adding a
CRLF line ending, and claims that 1 file needed sorting, even though
the file is left unchanged otherwise.

No idea why, but in any case the problem can be avoided by moving the
comment to before the package declaration.