Simplify SshClient
With the HostConfigEntry available on the session from the moment it is
created, we don't need to pass along so many data through the internal
methods of SshClient. We get a single point once the IoConnectFuture is
done and has created the session where we can read data from the
HostConfigEntry and configure the session.

Split the actual session configuration in several overrideable methods
so that user code can customize it easily.

Remove no longer needed AttributeKeys.