Session creation: lazy initialize ClientUserAuthService
This service gets created when the session is created. At that time the
properties may not be set yet on the session, so the constructor cannot
read them. Read them later when the service is started.

Arguably creating the service when the session is created is too early,
it would be sufficient to create it once the SERVICE_REQUEST message
arrives. But that would be a more risky change.