Keep track of already known host keys for a session
Track verified host keys for a session, and if a host key in a re-KEX
matches one of these already known-to-be-good keys, accept it silently
without calling the ServerKeyVerifier.

If a re-KEX occurs during a session, it's basically possible that the
server sends a different host key than in the last one.