Finish the "hostkeys-00@openssh.com" host key rotation implementation
Implement the client side as far as registering the host keys in the
client session: send the signature challenge, and handle the reply.
Once the keys are verified, invoke a new NewHostKeysHandler on the
SshClient. (It gets the session as parameter so that it can figure out
which server sent the keys. But the handler is installed on the client
because if it updates the known_hosts file, that could affect many
sessions.)

Take care to handle certificates and RSA keys. Fix the server side to
correctly deal with these, too.

Updating the known_hosts file is not done; user code can install its
own NewHostKeysHandler on the SshClient to do so.

Remove the nested OpenSshCertificate.CertificateOption class; these
options are better handled as Map<String, String>. Remove unnecessary
generics in arguments: replace Collection<? extends PublicKey> by
Collection<PublicKey>.