GH-622: Improve HostConfigEntry
Parse quoted values properly following the OpenSSH rules. Unify the
property handling; distinguish properly between simple keys, list-valued
keys, and additive keys.

A token is any sequence of characters not containing blanks unless the
blanks are in a quoted part. A quoted part is enclosed in single or
double quotes. Token values are de-quoted when read. They may contain
comma-separated lists of items.

A simple key has one token as value; when resolving multiple
HostConfigEntries, the first match wins.

The value of a list key can be one or more tokens, separated by unquoted
blanks; the first match wins. An example is UserKnownHostsFile.

The value of an additive key is typically one token. When the key
appears multiple times, all values are accumulated. An example is
IdentityFile.

And there are two additive list keys: SendEnv and SetEnv.

HostConfigEntry.getProperty() is for simple keys. For list keys or
additive keys, it returns only the first value.

HostConfigEntry.getValues() is new and always returns all values as a
list.

The separate storage for hotstName, userName, port, and proxyJump has
been removed; these are backed now by the corresponding properties.