GH-740: Remove SecurityUtils.isECCSupported()
EC keys and key exchanges and the three standard curves used in SSH are
supported on all Java versions that Apache MINA sshd supports.

Support for EC keys in Java came in Java 1.5,[1][2] but the "EC"
algorithm name for KeyPairGenerator is mentioned only in the Java 6
documentation.[3] In any case, as of Java 8, which is the minimum
requirement for Apache MINA sshd, EC is fully supported by the standard
security providers.

EC is also fully supported by Bouncy Castle.

[1] https://docs.oracle.com/javase/1.5.0/docs/guide/security/CryptoSpec.html#AppA
[2] https://docs.oracle.com/javase/1.5.0/docs/guide/security/jce/JCERefGuide.html#AppA
[3] https://docs.oracle.com/javase/6/docs/technotes/guides/security/StandardNames.html