GH-751: Fix "long name" for SFTP v3
The "long name" in a SFTP v3 SSH_FXP_NAME record as returned when
listing a directory is supposed to contain some representation suitable
for showing to the user directly; the draft RFC recommends a format
similar to Linux "ls -al", and Apache MINA sshd implements that strictly
as proposed in the draft RFC.

However, there were a few problems.

First, while the "long name" would contain a size also for directory
entries, the corresponding SFTP attributes always had size zero. Fix
this by reporting the size in the attributes not only for files but
simply for all entries, including directories.

Second, when an SFTP server itself was based on an SftpFileSystem
(chaining of SFTP servers, which is not ideal, but is possible), the
owner info was lost because the server uses a standard DirectoryStream.
Fix this by propagating the upstream "long name" through the cached
attributes, and using it, if available, when generating the server's
own "long name" to send to the client. (So the server passes through
the "long name" from the upstream server, if there is one.)

To fix this also for the "." and ".." entries, which a DirectoryStream
normally does not report but which we _do_ get from the upstream SFTP
server, add a mechanism to create a not-quite-standard DirectoryStream
that _does_ include these entries. That way, the server has access to
the attributes cached from the upstream server, and can report correct
ownership info also for these entries.